src/vertex_ai/requirements.txt
python-dotenv

# stable diffusion
diffusers
accelerate
opencv-python

# vertex ai model serving
fastapi
google-cloud-storage